worker/requirements.txt
azure-cosmos
azure-storage-queue
python-dotenv
